requirements.txt
streamlit>=1.28.0
Pillow>=10.0.0
fpdf2>=2.7.6
PyPDF2>=3.0.0